- add network client interface and its implementation - refactor network client in balance repository
