+++ wallet_data/src/commonMain/kotlin/net/thechance/mena/wallet/data/extension/NetworkClient.kt
@@ -1,0 +1,27 @@
+package net.thechance.mena.wallet.data.extension
+
+import io.ktor.client.request.HttpRequestBuilder
+import io.ktor.client.statement.HttpResponse
+
+
+interface NetworkClient {
+    suspend fun get(
+        urlString: String,
+        block: HttpRequestBuilder.() -> Unit = {}
+    ): HttpResponse
+
+    suspend fun post(
+        urlString: String,
+        block: HttpRequestBuilder.() -> Unit = {}
+    ): HttpResponse
+
+    suspend fun put(
+        urlString: String,
+        block: HttpRequestBuilder.() -> Unit = {}
+    ): HttpResponse
+
+    suspend fun delete(
+        urlString: String,
+        block: HttpRequestBuilder.() -> Unit = {}
+    ): HttpResponse
+}
