+++ wallet_data/src/commonMain/kotlin/net/thechance/mena/wallet/data/repository/balance/BalanceRepositoryImpl.kt
@@ -1,8 +1,8 @@
 package net.thechance.mena.wallet.data.repository.balance
 
-import net.thechance.mena.wallet.data.exceptions.safeApiCall
 import net.thechance.mena.wallet.data.dto.BalanceDto
-import net.thechance.mena.wallet.data.exceptions.NetworkClient
+import net.thechance.mena.wallet.data.exceptions.safeApiCall
+import net.thechance.mena.wallet.data.extension.NetworkClient
 import net.thechance.mena.wallet.domain.repository.BalanceRepository
 import org.koin.core.annotation.Single
